pyproject.toml
[project]
name = "deep-research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "fastapi>=0.135.1",
    "langchain>=1.2.12",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.1.2",
    "newsapi-python>=0.2.7",
    "numpy>=2.4.3",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.23",
    "uvicorn>=0.42.0",
]
